+++ internal/super/project.go
@@ -20,6 +20,7 @@ package super
 
 import (
 	"context"
+	"github.com/onflow/flow-cli/internal/util"
 	"github.com/onflow/flow-go-sdk"
 	"github.com/onflow/flow-go-sdk/crypto"
 	"github.com/pkg/errors"
@@ -27,7 +28,6 @@ import (
 	"github.com/onflow/flow-cli/pkg/flowkit"
 	"github.com/onflow/flow-cli/pkg/flowkit/config"
 	flowkitProject "github.com/onflow/flow-cli/pkg/flowkit/project"
-	"github.com/onflow/flow-cli/pkg/flowkit/util"
 )
 
 var emulator = config.EmulatorNetwork.Name
